replace most BIT macros with a Bit() template for better 64-bit safety test_bits: add tests for all routines in bits.h hpet: add safety check unit.h, bits.h, DOMEvent.h: change T(~0) to ~T(0) (required when T is 64-bit)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6335 3db68df2-c116-0410-a063-a993310a9797

+++ source/scripting/DOMEvent.h
@@ -126,7 +126,7 @@ public:
 	void StopPropagation( JSContext* cx, uintN argc, jsval* argv );
 
 public:
-	CScriptEvent( const CStrW& Type, size_t TypeCode = ~0, bool Cancelable = true, bool Blockable = true );
+	CScriptEvent( const CStrW& Type, size_t TypeCode = ~size_t(0), bool Cancelable = true, bool Blockable = true );
 	static void ScriptingInit();
 };
